Don't limit EBS volume size of prestoclient
By default EC2 uses the size of the snapshot backing the AMI.

Limiting the EBS volume size causes the process to fail if the source
AMI is bigger.

In the case of prestoclient, the source AMI is the output of the presto
build. The presto build filter is generating an AMI of 20GB.