Download Zeppelin without verbose wget output
The download progress bar floods the Packer output log with thousands of
lines. It's not very useful and it quickly exhausts the terminal buffer,
which makes debugging more difficult or impossible.

Verbose output supression results in one neat line of output when the
file is downloaded:

```text
==> amazon-ebs: 2020-10-20 15:36:47 URL:https://downloads.apache.org/zeppelin/zeppelin-0.8.2/zeppelin-0.8.2-bin-all.tgz [998508624/998508624] -> "zeppelin-0.8.2-bin-all.tgz" [1]
```

I used this instead of quiet mode because I still want to see errors
when they happen.