Use general Ubuntu AMI, not EKS optimized version
Using the full name prefix ensures that the general AMI is selected and
not any other customized version.

The previous filter was selecting the EKS optimized version, but that could
differ depending on when Canonical publishes the images. The latest
image published of any version would win.

The general AMIs are listed here: https://cloud-images.ubuntu.com/locator/ec2/

The EKS optimized AMIs are listed here: https://cloud-images.ubuntu.com/docs/aws/eks/

"These images are customised specifically for the EKS service, and are
not intended as general OS images."