Deploy all presto nodes in same AZ
An earlier change to the subnets meant that not all the worker nodes were being
being deployed in the same subnet.

This change makes it explicit that the ASG creates all worker nodes in
the same AZ as the coordindator node.

Update the documentation on AZs to clarify that load balancers need more
than one AZ to be deployed.