Changes to run presto in my sandbox account (#22)
* Fix typo

* Fix presto variables file name

The build command would fail with the following error message:

```text
"presto.packer.json": stat presto.packer.json: no such file or directory
```

* Document default region for Packer

* Explicitly use public IP for SSH in Packer

For some reason it was timing out while using the private IP, even
though Packer's documented default behavior is to use the public IP.

* https://www.packer.io/docs/builders/amazon/ebs
* https://github.com/hashicorp/packer/issues/5627

* Explicitly use public IP for Packer prestoclient

* Don't limit EBS volume size of prestoclient

By default EC2 uses the size of the snapshot backing the AMI.

Limiting the EBS volume size causes the process to fail if the source
AMI is bigger.

In the case of prestoclient, the source AMI is the output of the presto
build. The presto build filter is generating an AMI of 20GB.

* Download Zeppelin without verbose wget output

The download progress bar floods the Packer output log with thousands of
lines. It's not very useful and it quickly exhausts the terminal buffer,
which makes debugging more difficult or impossible.

Verbose output supression results in one neat line of output when the
file is downloaded:

```text
==> amazon-ebs: 2020-10-20 15:36:47 URL:https://downloads.apache.org/zeppelin/zeppelin-0.8.2/zeppelin-0.8.2-bin-all.tgz [998508624/998508624] -> "zeppelin-0.8.2-bin-all.tgz" [1]
```

I used this instead of quiet mode because I still want to see errors
when they happen.

* Update deprecated type name syntax

* Add variable overrides for my env

* Use general Ubuntu AMI, not EKS optimized version

Using the full name prefix ensures that the general AMI is selected and
not any other customized version.

The previous filter was selecting the EKS optimized version, but that could
differ depending on when Canonical publishes the images. The latest
image published of any version would win.

The general AMIs are listed here: https://cloud-images.ubuntu.com/locator/ec2/

The EKS optimized AMIs are listed here: https://cloud-images.ubuntu.com/docs/aws/eks/

"These images are customised specifically for the EKS service, and are
not intended as general OS images."

* Remove deprecated Packer spot option

The option spot_price_auto_product is no longer necessary and will cause
future versions of Packer to fail.

* Revert "Don't limit EBS volume size of prestoclient"

This reverts commit 45e085fb00dc5da53ef4fdb308beab0864558d34.

This section is actually required to ensure enough space for Zeppelin
when using an 8GB base AMI.

* Link to Canonical-maintained Ubuntu images

* Revert "Add variable overrides for my env"

This reverts commit a4b545e000975faec7a55054cb1b62c4dfb9e203.

TODO: Use tfvars file instead.

* Refactor subnets

* Make example query copy-pastable

* Create coordinator in same AZ as its EBS volume

Otherwise the coordinator will fail to set itself up upon failing to
mount the volume.

* Remove comments added during debugging

These were not intended to be added to the main code base.

The comment about EBS mount loop behavior needs more discussion to
decide how to handle it.

* Deploy all presto nodes in same AZ

An earlier change to the subnets meant that not all the worker nodes were being
being deployed in the same subnet.

This change makes it explicit that the ASG creates all worker nodes in
the same AZ as the coordindator node.

Update the documentation on AZs to clarify that load balancers need more
than one AZ to be deployed.

* Improve Terraform documentation

* Declare the prestoclient AMI as a dependency
* Document the new subnet_ids variable
* Clarify the preferred way to override variables
* Clarify the requirement to add a client node for UI access

* Move tfvars phrase to Terraform docs